# Conflicts:
#	src/main/java/com/example/yangyistarter/controller/UserController.java
#	src/main/java/com/example/yangyistarter/repository/UserRepository.java
#	src/main/java/com/example/yangyistarter/service/ParkingLotService.java
#	src/main/java/com/example/yangyistarter/service/UserService.java
#	src/test/java/com/example/yangyistarter/service/UserServiceTest.java

+++ src/main/java/com/example/yangyistarter/controller/PickingController.java
@@ -7,15 +7,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
-
 @RestController
 @AllArgsConstructor
 public class PickingController {
     PickingService pickingService;
     @PostMapping("/cars")
-    public ResponseCode picking(@RequestParam BigInteger receiptId) {
-
+    public ResponseCode picking(@RequestParam Long receiptId) {
         return pickingService.picking(receiptId);
     }
 }
