+++ src/main/java/com/example/yangyistarter/repository/ParkingSpaceRepository.java
@@ -4,9 +4,6 @@ import com.example.yangyistarter.entity.ParkingSpace;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface ParkingSpaceRepository extends JpaRepository<ParkingSpace, BigInteger> {
-
+public interface ParkingSpaceRepository extends JpaRepository<ParkingSpace, Long> {
 }
